--- C言語/file.txt
@@ -1,2 +1,0 @@
-miku   39/voc   aaa
-rin    393/guit   bbb
