+++ C言語/memo/C_memo.c
@@ -1,0 +1,13 @@
+/*
+C言語メモ
+http://gabekore.org/c-lang-dev-environment#i-3
+
+gcc -o test test.c
+test.exe
+
+or
+
+make
+hello
+
+*/
