+++ C言語/memo/file.txt
@@ -1,0 +1,2 @@
+miku   39/voc   aaa
+rin    393/guit   bbb
