+++ C言語/memo/memo1.c
@@ -1,0 +1,66 @@
+/*
+文字型 char
+整数型 int
+浮動小数点型 float double
+型なし void
+
+const修飾子-->定数になる
+*/
+
+/*
+静的変数について
+static
+
+関数の中で局所変数を宣言する際に記憶クラスとしてstaticを与えると、
+その関数が終了しても消滅せず、値を残すことができます。
+staticな局所変数は定義時に0で初期化されます。
+普通の局所変数のように関数がコールされた時のみ存在するのではなく、永久に存在します。
+しかし、外部変数とは異なり、他の関数からは利用できません。
+*/
+
+
+/*
+論理演算子
+||または
+&&かつ
+|否定
+*/
+
+
+/*
+条件演算子 ?
+[条件] ? [TRUEの場合の返却値] : [FALSEの場合の返却値] ;
+*/
+
+
+/*
+pointa演算子
+* : ポインタの指しているアドレスの内容をとりだす
+& : 変数の格納されているアドレスをとりだす
+
+*/
+
+
+/*
+キャスト演算子
+式のデータ型を()の中で指定した型に変換します。
+具体的には次のように記述します。
+
+double a = 10.5;
+int b = (int)a;
+
+上記の例では変数「a」の値を「int」型に変換したものを
+左辺の変数「b」に格納します。
+*/
+
+
+/*
+出力フォーマット指定子
+
+printf()などで使用する指定子である。
+指定子	対応する型	説明	使用例
+%c	char	１文字を出力する	"%c"
+%s	char *	文字列を出力する	"%8s", "%-10s"
+%d	int, short	整数を10進で出力する
+
+*/
