--- C言語/memo4.c
@@ -1,54 +1,0 @@
-//構造体
-
-/*
-幾つかの異なる型のデータをまとめて 1つのデータ型として扱うもの
-
-（１）構造体の型枠の宣言
-
-まず、複数のデータ（メンバと呼びます）をまとめて１つの構造体の型枠を宣言します。
-この宣言は単に構造体の型を作ったにすぎず、領域の割当ては行われていません。
-
-（書き方）
-
-struct	タグ名 {
-	データ型　メンバ名;
-};
-（例）
-
-struct seiseki {
-	int	no;
-	char	name[20];
-	double	average;
-};
-
-
-（４）構造体の参照
-
-構造体の各メンバは、「構造体変数名. メンバ名」のようにピリオド（.）を用いて指定します。
-この（.）を「ドット演算子」と呼びます。
-
-構造体変数の参照
-printf("%d %s %5.1f\n\n", seito1.no, seito1.name, seito1.average);
-
-構造体配列の参照
-for(i = 0; i < 3; i++) {
-	printf("%d %s %5.1f\n",
-		seito2[i].no, seito2[i].name, seito2[i].average);
-}
-
-
-関数のプロトタイプ宣言 : https://goo.gl/9T3Uef
-
-関数を呼び出す場合
-「呼び出す関数は使用する前に記述しておく必要がある」
-というルールがあります。
-
-しかし、関数が増えてくるとこれらの呼び出し順番を管理するのがめんどくさくなります。
-
-そんなときのために？C言語には「プロトタイプ宣言」という構文が用意されています。
-これは関数の中身（処理）を記述せずにとりあえず関数を呼び出す前に宣言しておいて、
-処理の部分は後ほど記述するという方法です。
-これで、いつでも関数を利用できるようになります。
-
-
-*/
