+++ handy/0508_memo.txt
@@ -1,0 +1,65 @@
+・tnsnames.ora
+
+C:\app\Administrator\product\11.2.0\client_1\network\admin
+
+・所在地：https://goo.gl/VuH3yF
+
+・あげお駅ホテル
+・->北本駅からバス(ワコーレ北本停留所下車)
+https://goo.gl/ectiw8
+https://goo.gl/1l7l5U
+
+
+・服装：カジュアル
+・カッターナイフ
+・メモ
+
+・日報の確認->瀧見さんに
+
+
+
+環境変数
+C:\app\Administrator\product\11.2.0\client_1\precomp\public
+
+C:\app\Administrator\product\11.2.0\client_1\precomp\LIB
+C:\app\Administrator\product\11.2.0\client_1\precomp\LIB\msvc
+
+C:\ryutu\bin
+C:\ryutu\delphi;C:\app\Administrator\product\11.2.0\client_1\BIN;C:\Program Files\Microsoft Office\Office14;C:\Program Files\Microsoft Office 15\root\office15
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+*/
