+++ handy/0508_ques.txt
@@ -1,0 +1,107 @@
+00:物流とは
+小売店が商品をメーカーに発注する際、物流倉庫業者を介する
+ことで適切且つ迅速な在庫管理を行うことが出来るその流れ。
+
+物流センターは、倉庫で商品の在庫管理し、
+仕入先と得意先の間をつなぐ役割
+
+受注センター：出荷指示(物流センターへの)
+仕入れセンター：発注指示(メーカーへの)
+
+04:入庫、棚入れ
+物流倉庫に届いた商品を、倉庫内の保管場所へと収めることを入庫という
+
+通常入庫：仕入れセンター→メーカー→物流センター
+振替入庫：仕入れセンター→他センター→物流センター
+補償入庫：メーカー→営業所→物流センター
+
+
+05:出庫関連(引当)
+引当：受注数に対して出庫可能な在庫を仮に割り当てること
+→現在の在庫で、どれだけの出庫が可能か
+→欠品している商品と数量の把握
+
+方法：先取り、優先順、均等割付、受注数比例順
+
+マスタ：
+出荷の仕組み作り（標準化）を行う土台となるのが「商品マスタ管理」
+「商品コード」をまとめて管理するデータが「商品マスタ」
+企業、得意先、エリア
+
+？？？
+二重起動抑止の流れが分からない
+
+
+06:出庫関連(計画)
+計画：
+引当を行った受注を基に、商品の荷姿に対して作業データを作成すること
+
+
+つまり、受注受信後、直ぐに出庫計画を立てられるわけではなく
+・取り込みされた受注が正常であるか
+・企業別に出庫に必要な設定は整っているか
+の確認が必要。
+
+
+06:出庫関連(ピッキング)
+
+JANコード：「どの事業者の、どの商品か」を表す、世界共通の商品識別番号
+
+1:納品伝票発行
+2:車積案内表→ピッキングエリアごとでの個数口の把握に使用
+3:ITFラベル一括発行→集合包装用商品コードをバーコードシンボルで表示する場合
+に国際標準化されている14桁のバーコードシンボル
+4:ベンダーJAN貼り付け
+5:ケース出荷合計リスト
+
+----------------------------------------------
+？？？
+二重起動抑止の流れが分からない
+
+？？？
+04:入庫、棚入れ
+物流倉庫に届いた商品を、倉庫内の保管場所へと収めることを入庫という
+
+通常入庫：仕入れセンター→メーカー→物流センター
+振替入庫：仕入れセンター→他センター→物流センター
+補償入庫：メーカー→営業所→物流センター
+
+このメニューの開き方不明
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+*/
